pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circuit-motifs"
version = "0.1.0"
description = "Network motif analysis of LLM attribution graphs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Michael Kenney" },
]
keywords = [
    "mechanistic-interpretability",
    "network-motifs",
    "attribution-graphs",
    "circuit-tracing",
    "LLM",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-igraph>=0.11",
    "networkx>=3.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "requests>=2.31",
    "tqdm>=4.66",
]

[project.optional-dependencies]
app = [
    "streamlit>=1.30",
    "plotly>=5.18",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.urls]
Homepage = "https://github.com/mkenney2/circuit-motifs"
Repository = "https://github.com/mkenney2/circuit-motifs"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
